Add thread cancellation mechanism
ChibiOS does not allow a thread to end, it just puts it in suspend
forever. The pthread port actively uses the port context data structure
which is located in the thread working area. As long as the thread
exists, the synchronization objects in there are in use. This causes a
problem when another ChibiOS thread is created with an overlapping
working area (which the user should be able to do, as the previous
thread has 'ended'). Thus, it's necessary to actually cancel the thread
that has exited.